requirements.txt
Django==1.7
dj-database-url
dj-static
django-toolbelt
gunicorn
psycopg2
static3
wsgiref
django-grappelli
django-endless-pagination